Move VectorIcon from openide.awt to openide.util.ui.
This will put VectorIcon in the same module as ImageUtilities. ImageUtilities
might at some point need to make use of VectorIcon, for other HiDPI-related
improvements. Note that openide.awt depends on openide.util.ui (not the other
way around).
